Big refactoring: migration to riverpod
Finally get rid of GlobalData and InheritedWidget :)
Introduce several providers for different areas of global state.
Refactoring some stuff in the settings page.
This should now introduce a clean architecture. However this is not
finished yet, tests are not updated and still failing and some
functionality of the settings page is also not yet restored.