Remove download_$langCode from SharedPrefs
Don't have information on whether a language is/should be downloaded in
the SharedPreferences anymore as this is redundant and can lead to
different edge cases. Instead the app only looks at the device storage
now: Is the language downloaded in its folder or not.

That also means that the app now never starts a download on startup -
now all downloads are transparently done by the user.