Adjust main navigation when on translated pages
When viewing a translated page, the main menu shows icons indicating
whether other worksheets are available in that language or not. However
these icons open a dialog but aren't direct links anymore. Instead the
menu entries open the page in that language (fall back to the app
language if the translation is not available).
Tests are still missing