Increasing build number
First I thought I don't need a build number so I omitted it in
pubspec.yaml. However the build number is used as versionCode in
Android. When trying to upload the appbundle to the google developer
console to release this version, I got the error message
"app bundle Version code 1 has already been used. Try another version
code". This is the fix and for the future remember: Always increase the
build number for a version that is to be released to play store.