Refactor: One TestLanguageController for tests
There was a confusing number of TestLanguageController and
DummyLanguageController classes to simulate different functionality of
languageProvider in testing. Now there is exactly one
TestLanguageController that is fine for most test cases with some
essential optional parameters for different behavior.
In two cases there are now CustomTestLanguageController to simulate
more special behavior.