Make sure onboarding can be interrupted any time
In case the user closes the app during the onboarding: Make sure next
time they open the app they'll continue at the same place.
Remove the passOnboarding function: Makes code more readable. The cost
is a few ms of showing a loading indicator on the very first app start
but that's negligible